Remove sourcing of virtualenvwrapper in bashrc
There were some local configuration issues which made it difficult to
source virtualenvwrapper in .bashrc because it required a
machine-specific rc file to be loaded which, per the current system, is
only set to be loaded AFTER the bashrc file was loaded.

This suggests a possible new feature a rc.first and rc.last system where
some local files are loaded first and some local files are loaded after.