Prevent tmux and path_helper from trashing PATH
When TMUX loads, it automatically loads `/etc/profile` prior to loading
`.bash_profile` which, on OSX
also runs path_helper. The effect is that if the PATH is already defined
and contains non-standard path directories they are rearranged, possibly
messing up your system. We now set the PATH to be empty and then
resource `/etc/profile` at the top of our `.bash_profile` so our PATH
can be fixed.

See the link below for a superuser answer relating to this issue.

https://superuser.com/questions/544989/does-tmux-sort-the-path-variable